Helper methods for pre-compilation features were moved from the `JavaScriptEngineUtils.cs` file to a separate file
